requirements.txt
click==7.1.2
dateutils==0.6.8
google-api-core==1.20.0
google-auth==1.16.1
google-cloud-bigquery==1.25.0
google-cloud-core==1.3.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
python-dateutil==2.8.1
requests==2.23.0
